(Bug 4775) Don't show promo text on receipts
Do show it when confirming a check/money order.

Ideally, we could check if the promo was active when the cart payment
was processed, but that's not an easy thing right now. So instead of
showing it always on all receipts when the promo period is active, just
don't show it on any receipt at all.