Applies labels to github issues
We look for labels in the format of "##name" in the comments of newly
opened issues and existing issues.

A label / list of labels in a comment will replace all existing labels.

Pull requests are automatically tagged "is: pull request". The comment
when opening a pull request is also checked for labels (but in practice,
probably not something that's going to be used that often)

Valid labels are:

* ##from: suggestions/support
* ##is: bug/feature
* ##effort: average/average/higher
* ##severity: minor/major/critical

Labels not on that list won't be applied.

Fixes dreamwidth/dw-free#701